feature: adjust the cli directory structure
