+++ cli/cmd/destroy.go
@@ -1,12 +1,13 @@
-package main
+package cmd
 
 import (
-	"github.com/spf13/cobra"
-	"github.com/chaosblade-io/chaosblade/exec"
-	"strings"
-	"github.com/chaosblade-io/chaosblade/transport"
-	"fmt"
 	"context"
+	"fmt"
+	"strings"
+
+	"github.com/chaosblade-io/chaosblade/exec"
+	"github.com/chaosblade-io/chaosblade/transport"
+	"github.com/spf13/cobra"
 )
 
 type DestroyCommand struct {
