+++ cli/cmd/exp.go
@@ -1,4 +1,4 @@
-package main
+package cmd
 
 import (
 	"context"
