+++ cli/cmd/query_jvm.go
@@ -1,9 +1,10 @@
-package main
+package cmd
 
 import (
-	"github.com/spf13/cobra"
 	"fmt"
+
 	"github.com/chaosblade-io/chaosblade/exec/jvm"
+	"github.com/spf13/cobra"
 )
 
 type QueryJvmCommand struct {
