+++ cli/cmd/version.go
@@ -1,8 +1,8 @@
-package main
+package cmd
 
 import (
-	"github.com/spf13/cobra"
 	"github.com/chaosblade-io/chaosblade/version"
+	"github.com/spf13/cobra"
 )
 
 var (
